system_backups/phoenix_backups/snapshot_20250816_152908/rag_system/requirements.txt
# Agent Exo-Suit V4.0 - RAG System Requirements
# Windows-compatible packages for GPU-accelerated RAG

# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Sentence transformers for embeddings
sentence-transformers>=2.2.0

# FAISS for vector search (CPU version for Windows compatibility)
faiss-cpu>=1.7.0

# NumPy for numerical operations
numpy>=1.21.0

# Additional utilities
tqdm>=4.64.0
scikit-learn>=1.1.0
transformers>=4.20.0

# Optional: GPU acceleration (if available)
# faiss-gpu>=1.7.0  # Uncomment if GPU FAISS is available on Windows
